refactor: extrair interfaces para arquivos dedicados
