+++ nexus-backend/src/modules/deliveries/interfaces/notification-payload.interface.ts
@@ -1,0 +1,13 @@
+import type { NotificationType } from './notification.interface';
+
+/**
+ * Interface para payload de notificação
+ */
+export interface NotificationPayload {
+  type: NotificationType;
+  deliveryId?: string;
+  userId?: string;
+  customerId?: string;
+  driverId?: string;
+  metadata?: Record<string, unknown>;
+}
