+++ nexus-backend/src/modules/deliveries/interfaces/request-user.interface.ts
@@ -1,0 +1,9 @@
+/**
+ * Interface para usuário autenticado na requisição
+ */
+export interface RequestUser {
+  id: string;
+  email: string;
+  name?: string;
+  roles?: string[];
+}
