+++ nexus-backend/src/modules/deliveries/interfaces/template-variables.interface.ts
@@ -1,0 +1,14 @@
+/**
+ * Interface para variáveis de template
+ */
+export interface TemplateVariables {
+  trackingCode: string;
+  customerName?: string;
+  driverId?: string;
+  driverName?: string;
+  estimatedTime?: string;
+  deliveryAddress?: string;
+  deliveryDate?: string;
+  statusMessage?: string;
+  [key: string]: string | undefined;
+}
