+++ nexus-backend/src/modules/deliveries/interfaces/tracking-code.interface.ts
@@ -1,0 +1,9 @@
+/**
+ * Interface para código de rastreamento parseado
+ */
+export interface ParsedTrackingCode {
+  prefix: string;
+  date: string;
+  sequence: string;
+  fullCode: string;
+}
