+++ nexus-backend/src/modules/deliveries/interfaces/validation-result.interface.ts
@@ -1,0 +1,8 @@
+/**
+ * Interface para resultado de validação
+ */
+export interface ValidationResult {
+  valid: boolean;
+  errors: string[];
+  warnings?: string[];
+}
